Code Refactor
- Break up monolithic functions into smaller pieces and separate files.
- Add more code comments.
- Combine duplicated logic into single `analyzeESLintReport` function.
